Rename directory to `bootstrap`
