+++ src/compiler/bootstrap/impl/ucc-processor.capability-activation.ts
