+++ src/compiler/bootstrap/ucc-schema-index.spec.ts
