+++ src/compiler/deserialization/bigint.ucrx.class.ts
@@ -1,7 +1,7 @@
 import { esline } from 'esgen';
 import { UcBigInt } from '../../schema/numeric/uc-bigint.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UC_MODULE_DESERIALIZER } from '../impl/uc-modules.js';
-import { UccConfig } from '../processor/ucc-config.js';
 import { UcrxCore } from '../rx/ucrx-core.js';
 import { UcrxLib } from '../rx/ucrx-lib.js';
 import { UcrxSetup } from '../rx/ucrx-setup.js';
