+++ src/compiler/deserialization/integer.ucrx.class.ts
@@ -1,7 +1,7 @@
 import { esline } from 'esgen';
 import { UcInteger } from '../../schema/numeric/uc-integer.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UC_MODULE_CHURI, UC_MODULE_DESERIALIZER } from '../impl/uc-modules.js';
-import { UccConfig } from '../processor/ucc-config.js';
 import { UcrxCore } from '../rx/ucrx-core.js';
 import { UcrxLib } from '../rx/ucrx-lib.js';
 import { UcrxSetup } from '../rx/ucrx-setup.js';
