+++ src/compiler/deserialization/ucd-process-defaults.ts
@@ -1,4 +1,4 @@
-import { UccConfig } from '../processor/ucc-config.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UcrxSetup } from '../rx/ucrx-setup.js';
 import { ucdProcessNonFinite } from './ucd-process-non-finite.js';
 import { ucdProcessPrimitives } from './ucd-process-primitives.js';
