+++ src/compiler/deserialization/ucd-process-primitives.ts
@@ -1,4 +1,4 @@
-import { UccConfig } from '../processor/ucc-config.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UcrxSetup } from '../rx/ucrx-setup.js';
 import { BigIntUcrxClass } from './bigint.ucrx.class.js';
 import { BooleanUcrxClass } from './boolean.ucrx.class.js';
