+++ src/compiler/mod.ts
@@ -5,9 +5,9 @@
  *
  * @module churi/compiler.js
  */
+export * from './bootstrap/mod.js';
 export * from './common/mod.js';
 export * from './deserialization/mod.js';
-export * from './processor/mod.js';
 export * from './rx/mod.js';
 export * from './serialization/mod.js';
 export * from './validation/mod.js';
