+++ src/compiler/serialization/ucs-process-defaults.ts
@@ -1,4 +1,4 @@
-import { UccConfig } from '../processor/ucc-config.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { ucsProcessPrimitives } from './ucs-process-primitives.js';
 import { UcsSetup } from './ucs-setup.js';
