+++ src/compiler/serialization/ucs-process-inset.ts
@@ -1,5 +1,5 @@
 import { UcFormatName } from '../../schema/uc-presentations.js';
-import { UccConfig } from '../processor/ucc-config.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UcsSetup } from './ucs-setup.js';
 
 export function ucsProcessInset(setup: UcsSetup): UccConfig<UcsInsetOptions> {
