+++ src/compiler/serialization/ucs-process-list.ts
@@ -4,10 +4,10 @@ import { ucModelName } from '../../schema/uc-model-name.js';
 import { ucNullable } from '../../schema/uc-nullable.js';
 import { ucOptional } from '../../schema/uc-optional.js';
 import { UcModel, UcSchema } from '../../schema/uc-schema.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UccListOptions } from '../common/ucc-list-options.js';
 import { UnsupportedUcSchemaError } from '../common/unsupported-uc-schema.error.js';
 import { UC_MODULE_SERIALIZER } from '../impl/uc-modules.js';
-import { UccConfig } from '../processor/ucc-config.js';
 import { ucsFormatCharge } from './impl/ucs-format-charge.js';
 import { UcsFormatterContext, UcsFormatterSignature } from './ucs-formatter.js';
 import { UcsSetup } from './ucs-setup.js';
