+++ src/compiler/serialization/ucs-process-number.ts
@@ -1,6 +1,6 @@
 import { UcNumber } from '../../schema/numeric/uc-number.js';
 import { UcSchema } from '../../schema/uc-schema.js';
-import { UccConfig } from '../processor/ucc-config.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { ucsFormatCharge } from './impl/ucs-format-charge.js';
 import { ucsFormatNumber } from './impl/ucs-format-number.js';
 import { UcsSetup } from './ucs-setup.js';
