+++ src/compiler/serialization/ucs-support-plain-text.ts
@@ -4,10 +4,10 @@ import { UcBigInt } from '../../schema/numeric/uc-bigint.js';
 import { UcString } from '../../schema/string/uc-string.js';
 import { ucModelName } from '../../schema/uc-model-name.js';
 import { ucsWriteAsIs } from '../../serializer/ucs-write-asis.js';
+import { UccCapability } from '../bootstrap/ucc-capability.js';
+import { UccConfig } from '../bootstrap/ucc-config.js';
 import { UnsupportedUcSchemaError } from '../common/unsupported-uc-schema.error.js';
 import { UC_MODULE_SERIALIZER } from '../impl/uc-modules.js';
-import { UccCapability } from '../processor/ucc-capability.js';
-import { UccConfig } from '../processor/ucc-config.js';
 import { ucsFormatBigInt } from './impl/ucs-format-bigint.js';
 import { ucsFormatBoolean } from './impl/ucs-format-boolean.js';
 import { ucsFormatInteger } from './impl/ucs-format-integer.js';
